Removing VLAN as an ether type.
Setting a packet ether type of VLAN/UDP, VLAN/TCP, does not seem to be very
useful and the tool does not actually initialize packets correctly when in
this mode so we are removing the command to avoid confusion.  Since the
ability to add a VLAN tag is available as a separate command it seems more
appropriate to limit the ether type commands to IPv4 and IPv6 only.